Begin move from standlone script to Django views.
Used bokeh's django_embed example as a template for this app, largely
structuring the same way, with a handler for each view and using the
autoload function (in config/urls.py) to serve each bokeh Document. That
solved the issue I thought I was going to have of having to manually
serve up each page with the bokeh serve command.

Initially I thought I would have been able to import the charts into the
views, but that throws an error - seems like it would allow for same
input/chart to be served up to multiple users. So I had to figure out a
way to try to keep the code concise as possible but yet allow the
interactions between the inputs and the charts. Basically created a
factory function to create the charts (in ghg_calc.py). The inputs and
have to go in each view function so they can update the data.