Put sequestration on bottom of stacked chart
Had to move it around in several places. Placing it at the bottom makes
if offset against the x axis, and bringing the whole bar down, as it
should. It gets hidden by the other values, but this is an improvement
over it being at the top and appearing to be positive.