Put bokeh apps into their own scripts
This massive reorganization was necessary because I was unable to find a
way to get bokeh to recognize the subpath of app/ghg for its static
assets. From some comments on bokeh's discourse and github, it seems
that the originators of the bokeh/django integration I followed are no
longer involved and that part of the code is not really being
maintained. So now instead of the bit of magic with the autoload
function that had been in config/urls.py creating each bokeh app, they
need to be created manually (with --prefix=app/ghg). Easy enough to do
with systemd service files and nginx locations, but much more code in
production (but at least it works).